requirements.txt
amqp==5.1.1
asgiref==3.7.2
astroid==2.15.5
billiard==4.1.0
celery==5.3.1
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.6
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
install==1.3.5
isort==5.12.0
kombu==5.3.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==10.0.0
platformdirs==3.8.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pylint==2.17.4
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
